requirements.txt
# ============================================================================
# ЗАВИСИМОСТИ ПРОЕКТА: Poizon → WordPress Сервис
# ============================================================================
# Этот файл содержит список всех необходимых Python библиотек для работы
# приложения синхронизации товаров Poizon → WordPress
# 
# Установка: pip install -r requirements.txt
# ============================================================================

# --- Веб-фреймворк ---
# Flask используется для создания веб-интерфейса и REST API endpoints
flask==3.0.0                    # Легковесный веб-фреймворк, обработка HTTP
flask-login==0.6.3              # Управление сессиями пользователей и авторизацией
werkzeug==3.0.1                 # Хеширование паролей для безопасности
gunicorn==21.2.0                # WSGI HTTP сервер для production (обязательно для деплоя!)

# --- Concurrency ---
# Gevent for high-performance networking and concurrency
gevent==24.2.1                  # Asynchronous networking library for Gunicorn worker

# --- Task Queue & Caching ---
# Celery for background task processing, Redis for broker and cache
celery==5.4.0                   # Distributed task queue
redis==5.0.4                    # Python client for Redis (broker and cache)

# --- Работа с HTTP и API ---
# Requests - основная библиотека для всех API запросов (Poizon, WooCommerce)
requests==2.31.0                # HTTP клиент для API запросов
urllib3==2.1.0                  # Low-level HTTP библиотека (зависимость requests)

# --- Обработка данных ---
# Pandas используется для манипуляции данными товаров и экспорта в CSV
pandas==2.3.3                   # Обработка табличных данных, DataFrame операции

# --- Обработка изображений ---
# Pillow для изменения размера изображений до 600x600 с сохранением пропорций
Pillow==10.4.0                  # Обработка и ресайз изображений товаров

# --- Парсинг контента ---
# BeautifulSoup для парсинга HTML описаний товаров (если потребуется)
beautifulsoup4==4.12.3          # Парсинг HTML/XML контента

# --- Конфигурация ---
# Python-dotenv для безопасного хранения API ключей в .env файле
python-dotenv==1.0.1            # Загрузка переменных окружения из .env

# --- Визуализация прогресса ---
# TQDM показывает прогресс-бары при длительных операциях загрузки
tqdm==4.67.1                    # Прогресс-бары в консоли и логах

# --- Сборка EXE (опционально) ---
pyinstaller==6.10.0             # Создание исполняемых файлов

# --- Инструменты разработки (опционально) ---
pytest==8.3.5                   # Тестирование
pytest-cov==6.0.0               # Покрытие тестами
black==24.8.0                   # Форматирование кода
flake8==7.1.1                   # Линтер
mypy==1.13.0                    # Проверка типов

# ============================================================================
# ПРИМЕЧАНИЕ:
# - Все версии зафиксированы для стабильной работы
# - Не меняйте версии без необходимости
# - При ошибках установки попробуйте: pip install --upgrade pip
# ============================================================================
